Ported changes from ZDoom revision 558-562: - Added skill definitions to MAPINFO. - Added text color options to menu code so that the skill definitions can set the color of the menu's text. - Externalized skill settings to ZDoom.pk3 - Changed names of skill properties to be the same as for episodes if possible. - minor fixes in custom skill code - Fixed: Backpacks didn't increase the ammo capacity to BackpackMaxAmount if an ammo type's BackpackAmount was 0. - Fixed some minor bugs in custom skill code.
SVN r982 (trunk)

+++ src/p_3dfloors.h
@@ -45,7 +45,7 @@ struct F3DFloor
 	struct planeref
 	{
 		secplane_t *	plane;
-		short *			texture;
+		int *			texture;
 		fixed_t *		texheight;
 		sector_t *		model;
 		bool			isceiling;
